moved mock file in frontend test
