+++ frontend/tests/components/scan/ScanForm/ScanForm.validation.test.jsx
@@ -7,14 +7,20 @@ import userEvent from "@testing-library/user-event";
 import ScanForm from "../../../../src/components/scan/ScanForm";
 import RecentScans from "../../../../src/components/scan/utils/RecentScans";
 
-import { mockedUseAuthStore, mockedUseTagsStore, mockedUsePluginConfigurationStore } from "./mock";
+import {
+  mockedUseAuthStore,
+  mockedUseTagsStore,
+  mockedUsePluginConfigurationStore,
+} from "../../../mock";
 
 jest.mock("axios");
 // IMPORTANT: this mocks work with several storages because all of them are imported from index!
 jest.mock("../../../../src/stores", () => ({
   useAuthStore: jest.fn((state) => state(mockedUseAuthStore)),
   useTagsStore: jest.fn((state) => state(mockedUseTagsStore)),
-  usePluginConfigurationStore: jest.fn((state) => state(mockedUsePluginConfigurationStore))
+  usePluginConfigurationStore: jest.fn((state) =>
+    state(mockedUsePluginConfigurationStore),
+  ),
 }));
 // mock RecentScans component
 jest.mock("../../../../src/components/scan/utils/RecentScans", () =>
